Netflix Dashbord PowerBI
A summary view showcasing the total number of TV shows and movies available on Netflix, categorized by genres and release years. This provides an instant understanding of the content library's composition.allowing users to analyze the popularity of different genres or content categories. This helps identify which types of content resonate most with the audience.Insights into user ratings and reviews for TV shows and movies. This helps assess content quality and understand viewer sentiment, aiding in decision-making for content acquisition or production.